+++ src/__tests__/sideEffects.test.js
@@ -9,7 +9,7 @@ import {
 } from 'rxjs/operators'
 import { PENDING, SUCCESS, FAILURE, CLEAN, RUN, CANCEL } from '../actionTypes'
 import { createTestRJSubscription } from '../testUtils'
-import {
+import RxEffects, {
   TAKE_EFFECT_EVERY,
   TAKE_EFFECT_GROUP_BY,
   // TAKE_EFFECT_QUEUE,
@@ -31,6 +31,56 @@ describe('RJ side effect model', () => {
 
     const RjObject = rj({
       effect: mockApi,
+    })
+
+    const subject = createTestRJSubscription(RjObject, mockCallback)
+
+    subject.next({
+      type: RUN,
+      payload: { params: [] },
+      meta: {},
+      callbacks: {},
+    })
+
+    mockApi.mock.results[0].value.then(() => {
+      expect(mockCallback).toBeCalledTimes(3)
+
+      expect(mockCallback).nthCalledWith(1, {
+        type: RUN,
+        payload: { params: [] },
+        meta: {},
+        callbacks: {},
+      })
+
+      expect(mockCallback).nthCalledWith(2, {
+        type: PENDING,
+        meta: {},
+      })
+
+      expect(mockCallback).nthCalledWith(3, {
+        type: SUCCESS,
+        meta: {},
+        payload: {
+          params: [],
+          data: mockApiResult,
+        },
+      })
+
+      done()
+    })
+  })
+
+  it('should support using rxEffects as function as valid take effects', (done) => {
+    const mockApiResult = [
+      { id: 1, name: 'Alice' },
+      { id: 2, name: 'Bob' },
+    ]
+    const mockApi = jest.fn().mockResolvedValueOnce(mockApiResult)
+    const mockCallback = jest.fn()
+
+    const RjObject = rj({
+      effect: mockApi,
+      takeEffect: RxEffects.latest,
     })
 
     const subject = createTestRJSubscription(RjObject, mockCallback)
@@ -1504,21 +1554,16 @@ describe('RJ side effect model', () => {
     const mockApi = jest.fn().mockResolvedValue(1312)
     const customMockTakeEffect = jest
       .fn()
-      .mockImplementation(
-        (
-          actionObservable,
-          stateObservable,
-          { effect, runSideEffectAction, getEffectCaller }
-        ) =>
-          forkJoin({
-            drago: from(effect()),
-            drago2x: from(effect()),
-          }).pipe(
-            map((result) => ({
-              type: '2X',
-              payload: result,
-            }))
-          )
+      .mockImplementation((actionObservable, stateObservable, { effect }) =>
+        forkJoin({
+          drago: from(effect()),
+          drago2x: from(effect()),
+        }).pipe(
+          map((result) => ({
+            type: '2X',
+            payload: result,
+          }))
+        )
       )
 
     const RjObject = rj({
@@ -1542,8 +1587,8 @@ describe('RJ side effect model', () => {
       expect.any(Observable),
       {
         effect: mockApi,
-        runSideEffectAction: expect.any(Function),
         getEffectCaller: expect.any(Function),
+        prefix: '',
       }
     )
 
@@ -1849,8 +1894,8 @@ describe('RJ side effect model', () => {
       expect.any(Observable),
       {
         effect: mockApi,
-        runSideEffectAction: expect.any(Function),
         getEffectCaller: expect.any(Function),
+        prefix: '',
       }
     )
 
@@ -1875,6 +1920,57 @@ describe('RJ side effect model', () => {
         onSuccess: undefined,
         onFailure: undefined,
       },
+    })
+  })
+
+  it('should hanlde custom side effects ... using core rj take effects', async () => {
+    const mockApi = jest.fn().mockResolvedValue('GioVa')
+
+    const mockCallback = jest.fn()
+
+    const RjObject = rj({
+      addSideEffect: (actions, state, config) =>
+        RxEffects.latest(actions, state, {
+          ...config,
+          prefix: 'GANG/',
+        }),
+      actions: () => ({
+        gang: (...params) => makeAction(`GANG/${RUN}`, ...params),
+      }),
+      effect: mockApi,
+    })
+
+    const { actionCreators } = RjObject
+    const subject = createTestRJSubscription(RjObject, mockCallback)
+    const dispatch = (action) => subject.next(action)
+    const actions = bindActionCreators(actionCreators, dispatch, subject)
+
+    actions.gang(23)
+
+    await mockApi.mock.results[0].value
+
+    expect(mockCallback).nthCalledWith(1, {
+      type: 'GANG/RUN',
+      payload: {
+        params: [23],
+      },
+      meta: {},
+      callbacks: {
+        onSuccess: undefined,
+        onFailure: undefined,
+      },
+    })
+    expect(mockCallback).nthCalledWith(2, {
+      type: 'GANG/PENDING',
+      meta: {},
+    })
+    expect(mockCallback).nthCalledWith(3, {
+      type: 'GANG/SUCCESS',
+      payload: {
+        params: [23],
+        data: 'GioVa',
+      },
+      meta: {},
     })
   })
 })