+++ src/index.js
@@ -42,3 +42,7 @@ export * from './types'
 
 // All rx effects
 export { default as RxEffects } from './rxEffects'
+
+// Helper for custom Rx effect:
+// { type: RUN } => [Observable<PENDING>, Observable<SUCCESS | FAILURE>]
+export { default as mapRunActionToObservable } from './mapRunActionToObservable'